+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 0.4.1.0 -- Unreleased
 
-* New built-in functions: `\interval`, `\read`, `\show`.
+* New built-in functions: `\interval`, `\read`, `\show`, `\tuple`.
+* New particle: `\fold1`.
 
 ## 0.4.0.0 -- 2023-07-10